[aoc2019] Support large numbers in intcode computer

+++ 2019/aoc07/src/main.rs
@@ -5,9 +5,9 @@ fn main() {
     let filename = "input/input.txt";
     let contents = fs::read_to_string(filename).expect("Something went wrong reading the file.");
 
-    let int_code_program: Vec<i32> = contents
+    let int_code_program: Vec<i64> = contents
         .split(",")
-        .map(|x| x.parse::<i32>().unwrap())
+        .map(|x| x.parse::<i64>().unwrap())
         .collect();
 
     let p = permutations(5).collect::<Vec<_>>();
@@ -17,23 +17,23 @@ fn main() {
         let mut amp_3 = IntCodeInterpreter::new(int_code_program.clone());
         let mut amp_4 = IntCodeInterpreter::new(int_code_program.clone());
         let mut amp_5 = IntCodeInterpreter::new(int_code_program.clone());
-        amp_1.add_custom_input(i[0] as i32);
+        amp_1.add_custom_input(i[0] as i64);
         amp_1.add_custom_input(0);
         amp_1.run();
 
-        amp_2.add_custom_input(i[1] as i32);
+        amp_2.add_custom_input(i[1] as i64);
         amp_2.add_custom_input(amp_1.get_last_output());
         amp_2.run();
 
-        amp_3.add_custom_input(i[2] as i32);
+        amp_3.add_custom_input(i[2] as i64);
         amp_3.add_custom_input(amp_2.get_last_output());
         amp_3.run();
 
-        amp_4.add_custom_input(i[3] as i32);
+        amp_4.add_custom_input(i[3] as i64);
         amp_4.add_custom_input(amp_3.get_last_output());
         amp_4.run();
 
-        amp_5.add_custom_input(i[4] as i32);
+        amp_5.add_custom_input(i[4] as i64);
         amp_5.add_custom_input(amp_4.get_last_output());
         amp_5.run();
         println!("{}", amp_5.get_last_output());
