+++ 2019/aoc09/src/main.rs
@@ -5,9 +5,9 @@ fn main() {
     let filename = "input/input.txt";
     let contents = fs::read_to_string(filename).expect("Something went wrong reading the file.");
 
-    let int_code_program: Vec<i32> = contents
+    let int_code_program: Vec<i64> = contents
         .split(",")
-        .map(|x| x.parse::<i32>().unwrap())
+        .map(|x| x.parse::<i64>().unwrap())
         .collect();
 
     // let mut interpreter = IntCodeInterpreter::new(int_code_program);
